% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_singleclearing.R
\name{calc_segments}
\alias{calc_segments}
\title{Extracts segment coordination from data.frame}
\usage{
calc_segments(df)
}
\arguments{
\item{df}{}
}
\value{
combined data point
}
\description{
Extracts segment coordination from data.frame
}
\author{
Friedrich Krebs
}
