% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_singleclearing.R
\name{calc_segments_twisted}
\alias{calc_segments_twisted}
\title{Extracts segment coordination from data.frame  for a twisted diagramm}
\usage{
calc_segments_twisted(df)
}
\arguments{
\item{df}{}
}
\value{
combined data point
}
\description{
Extracts segment coordination from data.frame  for a twisted diagramm
}
\author{
Friedrich Krebs

Sascha Holzhauer
}
