% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_clearing.R
\name{clearing_calc}
\alias{clearing_calc}
\title{Calculate the system marginal price (SMP) and the amount of energy to be traded}
\usage{
clearing_calc(bids_df, offers_df)
}
\arguments{
\item{bids_df}{holds the demand curve}

\item{offers_df}{holds the supply curve}
}
\value{
data.frame with clearing_price and quantity_traded
}
\description{
Calculate the system marginal price (SMP) and the amount of energy to be traded
}
\author{
Friedrich Krebs
}
