% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_clearing.R
\name{clearing_calc_demand}
\alias{clearing_calc_demand}
\title{Calculate demand curve based on market data frame and a start time of the delivery period}
\usage{
clearing_calc_demand(market_df, start_time)
}
\arguments{
\item{market_df}{}

\item{start_time}{}
}
\value{
data frame of cumulated bids ordered by decreasing price
}
\description{
Calculate demand curve based on market data frame and a start time of the delivery period
}
\author{
Friedrich Krebs
}
