% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_clearing.R
\name{clearing_calc_supply}
\alias{clearing_calc_supply}
\title{Calculate supply curve based on market data frame and a start time of the dellivery period}
\usage{
clearing_calc_supply(market_df, start_time)
}
\arguments{
\item{market_df}{}

\item{start_time}{}
}
\value{
data frame of cumulated offers ordered by increasing price
}
\description{
Calculate supply curve based on market data frame and a start time of the dellivery period
}
\author{
Friedrich Krebs
}
