% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_clearing.R
\name{clearing_find_price_requested}
\alias{clearing_find_price_requested}
\title{Find the price for an amount of energy in a data frame of stacked bids or offers}
\usage{
clearing_find_price_requested(energy_amount, df)
}
\arguments{
\item{energy_amount}{}

\item{df}{holds demand or supply curve}
}
\value{
NaN if there is no price found for energy_amount
}
\description{
Find the price for an amount of energy in a data frame of stacked bids or offers
}
\author{
Friedrich Krebs
}
