% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{create_dexpa}
\alias{create_dexpa}
\title{The given/default dexpa object is dublicated, and sim$id and dp$dbname are subsituted by the given ID.}
\usage{
create_dexpa(id, dexpa = dexR::param_getDefaultDexpa())
}
\arguments{
\item{id}{id}
}
\value{
dexpa
}
\description{
The given/default dexpa object is dublicated, and sim$id and dp$dbname are subsituted by the given ID.
}
\author{
Sascha Holzhauer
}
