% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{create_dexpas}
\alias{create_dexpas}
\title{Create a list of <code>dexpa</code> objects from vector of ids}
\usage{
create_dexpas(ids, dexpa = dexR::param_getDefaultDexpa())
}
\arguments{
\item{ids}{vector of ids}
}
\value{
list of dexpas
}
\description{
The given/default dexpa object is dublicated, and sim$id and dp$dbname are subsituted by one of the given IDs.
}
\author{
Sascha Holzhauer
}
