% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_report_create.R
\name{createFullReport}
\alias{createFullReport}
\title{Create the full report}
\usage{
createFullReport(
  dexpa,
  outputfile,
  rmdfile = "DEX_report_full.Rmd",
  output_dir = dexpa$dirs$output$reports
)
}
\arguments{
\item{dexpa}{parameter object}

\item{outputfile}{filename of resulting PDF file}

\item{output_dir}{destination folder}
}
\value{
PDF file
}
\description{
Create the full report
}
\author{
Sascha Holzhauer
}
