% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_report_create.R
\name{createSingleReport}
\alias{createSingleReport}
\title{Create a single run report (slides)}
\usage{
createSingleReport(dexpas, outputfile,
  rmdfile = "DEX_report_single_slides.Rmd",
  output_dir = dexpa$dirs$output$reports, clean = T, run_pandoc = T)
}
\arguments{
\item{outputfile}{filename of resulting PDF file}

\item{output_dir}{destination folder}

\item{dexpa}{parameter object}
}
\value{
PDF file
}
\description{
Create a single run report (slides)
}
\author{
Sascha Holzhauer
}
