% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{demo_prepare_db4figures}
\alias{demo_prepare_db4figures}
\title{Import dumpfiles into DB to use for sample figures}
\usage{
demo_prepare_db4figures()
}
\value{
list of dexpa objects
}
\description{
Import dumpfiles into DB to use for sample figures
}
\author{
Sascha Holzhauer
}
