% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexR-package.R
\docType{package}
\name{dexR}
\alias{dexR}
\title{DexR}
\description{
The package offers deployment and anlyses methods for DEX.
}
\details{
DEX R

\tabular{ll}{
Package \tab		dexR \cr
Model:  \tab		<Basic> \cr
Licence:\tab		GPL v3 \cr
Be aware of:\tab	-\cr
Version: \tab	0.6.7 \cr
Date	   	\tab 	2019-05-15 \cr
Changes	\tab	refactoring, comment \cr
}
}
\author{
Sascha Holzhauer, UniKassel \email{Sascha.Holzhauer@uni-kassel.de"}
}
\keyword{analyses}
\keyword{energy}
\keyword{market}
\keyword{simulation}
