% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{emggetport}
\alias{emggetport}
\title{Determine EMG port in multi-node (and other) setting}
\usage{
emggetport(dexpa, nodeid)
}
\arguments{
\item{dexpa}{}

\item{nodeid}{}
}
\value{
EMG port for specific node ID
}
\description{
Determine EMG port in multi-node (and other) setting
}
\author{
Sascha Holzhauer
}
