% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_clearingsingle.R
\name{hl_clearingsingle}
\alias{hl_clearingsingle}
\title{Gnerates a figure of clearing at starttime}
\usage{
hl_clearingsingle(
  dexpa,
  starttime = NULL,
  title = NULL,
  filename = "GetIntersectionPrice",
  dexpa_comp = NULL
)
}
\arguments{
\item{dexpa}{parameter list}

\item{starttime}{start time to consider in ms}

\item{title}{figure title}

\item{filename}{output filename}

\item{dexpa_comp}{parameter list of run to compare with}
}
\value{
figure in file
}
\description{
Gnerates a figure of clearing at starttime
}
\author{
Sascha Holzhauer
}
