% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_closeexperiment}
\alias{hl_closeexperiment}
\title{Stop EMGs, dump database, create full report, shutdown server, and redirect output.}
\usage{
hl_closeexperiment(
  dexpa,
  outputfile = "",
  basetime,
  offset = round(basetime - as.numeric(Sys.time()) * 1000),
  infoData = NULL,
  nodeids = c("")
)
}
\arguments{
\item{dexpa}{}
}
\value{

}
\description{
Stop EMGs, dump database, create full report, shutdown server, and redirect output.
}
\author{
Sascha Holzhauer
}
