% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_config.R
\name{hl_config_clearclients}
\alias{hl_config_clearclients}
\title{Deletes users from PostreDB (except admin, enavi, inspector)}
\usage{
hl_config_clearclients(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{

}
\description{
Deletes users from PostreDB (except admin, enavi, inspector)
}
\author{
Sascha Holzhauer
}
