% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_config.R
\name{hl_config_clients2db}
\alias{hl_config_clients2db}
\title{Stores client and role data from CSV file to PostGreSQL database.
The database tables are _not_ emptied before submission.}
\usage{
hl_config_clients2db(
  dexpa,
  sourcedir = paste(dexpa$dirs$config, dexpa$sim$id, sep = "/"),
  paramConfigs
)
}
\arguments{
\item{dexpa}{}

\item{sourcedir}{}

\item{sourcefile}{}
}
\value{
table in DB
}
\description{
Stores client and role data from CSV file to PostGreSQL database.
The database tables are _not_ emptied before submission.
}
\author{
Sascha Holzhauer
}
