% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_config.R
\name{hl_config_copycsvtemplates}
\alias{hl_config_copycsvtemplates}
\title{Copy CSV template to new configuration folder}
\usage{
hl_config_copycsvtemplates(dexpa, targetdir = paste(dexpa$dirs$config,
  dexpa$sim$id, sep = "/"))
}
\arguments{
\item{dexpa}{parameter list}

\item{targetdir}{only used to check existence of target directory}
}
\value{
CSV files
}
\description{
Copy CSV template to new configuration folder
}
\author{
Sascha Holzhauer
}
