% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_config.R
\name{hl_config_exportMarketProducts}
\alias{hl_config_exportMarketProducts}
\title{Export market products from DB to CSV file.}
\usage{
hl_config_exportMarketProducts(dexpa, dexpa_template,
  targetdir = paste(dexpa$dirs$config, dexpa$sim$id, sep = "/"))
}
\arguments{
\item{dexpa}{the new configuration (relevant for config folder to store CSV file in)}

\item{dexpa_template}{the template configuration (for the database)}

\item{targetdir}{(optional)}
}
\value{
CSV file
}
\description{
Export market products from DB to CSV file.
}
\author{
Sascha Holzhauer
}
