% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_config.R
\name{hl_config_marketProducts2db}
\alias{hl_config_marketProducts2db}
\title{Stores market products from CSV file to PostGreSQL database.
The database tables are emptied before submission.}
\usage{
hl_config_marketProducts2db(dexpa, sourcedir = paste(dexpa$dirs$config,
  dexpa$sim$id, sep = "/"),
  sourcefile = paste("DEX_Param_MarketProducts_", dexpa$sim$id, ".csv",
  sep = ""), firstDeliveryPeriodStart = Sys.time())
}
\arguments{
\item{dexpa}{\itemize{
    \item \code{dexpa$dirs$config}
    \item \code{dexpa$sim$id}
 \item \code{dexpa$sim$firstdeliverystart$delay}
 \item \code{dexpa$db$tablenames$marketproducts}
 \item \code{dexpa$db$tablenames$mmarketproducts}
}}

\item{sourcedir}{}

\item{sourcefile}{}

\item{firstDeliveryPeriodStart}{as POSIX object}
}
\value{
table in DB
}
\description{
Stores market products from CSV file to PostGreSQL database.
The database tables are emptied before submission.
}
\seealso{
\code{\link[dexr]{input_db_getconnection}}

\code{\link[RPostgreSQL]{dbWriteTable}}
}
\author{
Sascha Holzhauer
}
