% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment}
\alias{hl_experiment}
\title{Concuct experiment from starting backend server to creation of full report}
\usage{
hl_experiment(
  dexpa,
  basetime = as.numeric(round(Sys.time(), "mins")) * 1000,
  offset = round(basetime - as.numeric(Sys.time()) * 1000),
  outputfile = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id, "_",
    dexpa$emg$rseed, ".log", sep = ""),
  outfilemarket = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id,
    "_", dexpa$emg$rseed, "_market.log", sep = ""),
  outfileemg = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id, "_",
    dexpa$emg$rseed, "_emg.log", sep = ""),
  outfile,
  runmarket = T,
  runemg = T,
  shutdown = T,
  createdb = F,
  pauseafterxmlcreation = F
)
}
\arguments{
\item{dexpa}{}
}
\value{
report, database
}
\description{
Concuct experiment from starting backend server to creation of full report
}
\author{
Sascha Holzhauer
}
