% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_awaitemgstartup}
\alias{hl_experiment_awaitemgstartup}
\title{Sleep during EMG startup}
\usage{
hl_experiment_awaitemgstartup(dexpa, waittime = dexpa$emg$emgstartuptime)
}
\arguments{
\item{dexpa}{parameter list}

\item{waittime}{(default: dexpa$emg$emgstartuptime)}
}
\value{
-
}
\description{
Sleep during EMG startup
}
\author{
Sascha Holzhauer
}
