% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_bootbackend}
\alias{hl_experiment_bootbackend}
\title{Boots backend server (no configuration, no starting)}
\usage{
hl_experiment_bootbackend(dexpa, basetime, offset, outfilesys)
}
\arguments{
\item{dexpa}{}
}
\description{
Boots backend server (no configuration, no starting)
}
\author{
Sascha Holzhauer
}
