% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_configemg}
\alias{hl_experiment_configemg}
\title{Create configuration for EMGs to experiment}
\usage{
hl_experiment_configemg(dexpa, outfilesys = "")
}
\arguments{
\item{dexpa}{}
}
\value{
-
}
\description{
Create configuration for EMGs to experiment
}
\author{
Sascha Holzhauer
}
