% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_ensureFileLogging}
\alias{hl_experiment_ensureFileLogging}
\title{Ensures file exists an redirects output.}
\usage{
hl_experiment_ensureFileLogging(dexpa, outputfile)
}
\arguments{
\item{dexpa}{}

\item{outputfile}{}
}
\value{

}
\description{
Ensures file exists an redirects output.
}
\author{
Sascha Holzhauer
}
