% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_runbackend}
\alias{hl_experiment_runbackend}
\title{Run backend server to experiment}
\usage{
hl_experiment_runbackend(dexpa, outfilesys = "",
  basetime = as.numeric(round(Sys.time(), "mins")) * 1000,
  offset = round(basetime - as.numeric(Sys.time()) * 1000),
  startServer = TRUE)
}
\arguments{
\item{dexpa}{parameter object}

\item{outfilesys}{log file}

\item{basetime}{in ms, assigned to \code{de.unik.enavi.market.time.basetime}}

\item{offset}{in ms, assigned to \code{de.unik.enavi.market.time.offset}}

\item{startServer}{TRUE if the server should also be started}
}
\value{
-
}
\description{
Default for first delivery period start is the basetime.
}
\author{
Sascha Holzhauer
}
