% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_runemg}
\alias{hl_experiment_runemg}
\title{Run EMGs to experiment}
\usage{
hl_experiment_runemg(
  dexpa,
  runemg = T,
  outfileemg = "",
  outfilesys = "",
  pauseafterxmlcreation = F
)
}
\arguments{
\item{dexpa}{}
}
\value{
-
}
\description{
Run EMGs to experiment
}
\author{
Sascha Holzhauer
}
