% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_experiment_stopemg}
\alias{hl_experiment_stopemg}
\title{Stop EMG clients. Disables SSL verification.}
\usage{
hl_experiment_stopemg(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
-
}
\description{
Stop EMG clients. Disables SSL verification.
}
\author{
Sascha Holzhauer
}
