% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_clearing.R
\name{hl_figure_clearing_comp_clearingPriceByCTbyProduct}
\alias{hl_figure_clearing_comp_clearingPriceByCTbyProduct}
\title{Retrieve data and store figure about the clearing price per clearing by clearing time and by product}
\usage{
hl_figure_clearing_comp_clearingPriceByCTbyProduct(dexpas, ...)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieve data and store figure about the clearing price per clearing by clearing time and by product
}
\author{
Sascha Holzhauer
}
