% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_clearing.R
\name{hl_figure_clearing_numConsideredRequests}
\alias{hl_figure_clearing_numConsideredRequests}
\title{Retrieve data and store figure about the number of considered requests per clearing by clearing time}
\usage{
hl_figure_clearing_numConsideredRequests(dexpa, ...)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieve data and store figure about the number of considered requests per clearing by clearing time
}
\author{
Sascha Holzhauer
}
