% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_requested_byStatusByStartT}
\alias{hl_figure_energy_requested_byStatusByStartT}
\title{Retrieves requests data from DB and creates figure of the received ernergy per 
status by delivery start time.}
\usage{
hl_figure_energy_requested_byStatusByStartT(dexpa, skiplegend = F,
  type = "residual")
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of the received ernergy per 
status by delivery start time.
}
\author{
Sascha Holzhauer
}
