% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_requested_comp_sumByStartT}
\alias{hl_figure_energy_requested_comp_sumByStartT}
\title{Retrieves requests data from DB and creates figure of requested energy by delivery start time.}
\usage{
hl_figure_energy_requested_comp_sumByStartT(dexpas)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of requested energy by delivery start time.
}
\author{
Sascha Holzhauer
}
