% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_requested_comp_sumGenByTypeStartT}
\alias{hl_figure_energy_requested_comp_sumGenByTypeStartT}
\title{Retrieves requests data from DB and creates figure of requested energy from generation by generation type 
and delivery start time.}
\usage{
hl_figure_energy_requested_comp_sumGenByTypeStartT(dexpas,
  skiplegend = F)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of requested energy from generation by generation type 
and delivery start time.
}
\author{
Sascha Holzhauer
}
