% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_requested_comp_sumLoadGenByStartT}
\alias{hl_figure_energy_requested_comp_sumLoadGenByStartT}
\title{Retrieves requests data from DB and creates figure of requested energy by delivery start time 
(generation, load, and residual as lines).}
\usage{
hl_figure_energy_requested_comp_sumLoadGenByStartT(dexpas)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Considered stati for various product configurations:
Single product, single auction: ACCEPTED, PARTYL_ACCEPTED, DECLINED
Single product, multiple auctions: ACCEPTED, PARTYL_ACCEPTED (energy_accepted, last auction: energy_requested, DECLINED (last auction)
Multiple products: ACCEPTED, PARTYL_ACCEPTED (actually, PARTLY_ACCEPTED (energy_requested) and DECLINED need to be considered for the last
auction across products, which is not yet implemented).
}
\author{
Sascha Holzhauer
}
