% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_requested_sumLoadGenByStartT}
\alias{hl_figure_energy_requested_sumLoadGenByStartT}
\title{Retrieves requests data from DB and creates figure of requested energy by delivery start time 
(generation, load, and residual as lines).}
\usage{
hl_figure_energy_requested_sumLoadGenByStartT(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of requested energy by delivery start time 
(generation, load, and residual as lines).
}
\author{
Sascha Holzhauer
}
