% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_energy.R
\name{hl_figure_energy_status_comp_sumLoadGenByLocationAndStartT}
\alias{hl_figure_energy_status_comp_sumLoadGenByLocationAndStartT}
\title{Retrieves requests data from DB and creates figure of requested and accepted energy by delivery start time}
\usage{
hl_figure_energy_status_comp_sumLoadGenByLocationAndStartT(dexpas,
  stati = c("ALL_REQUESTED", "ACCEPTED", "PARTLY_ACCEPTED", "DECLINED"),
  ggplotaddons = NULL)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Facets: Location
Colour: RunID
LineType: Load/Generation
}
\details{
In case stati = ALL_REQUESTED, requests are filtered according to product configurations:
Single product, single auction: ACCEPTED, PARTYL_ACCEPTED, DECLINED
Single product, multiple auctions: ACCEPTED, PARTYL_ACCEPTED (energy_accepted, last auction: energy_requested, DECLINED (last auction)
Multiple products: ACCEPTED, PARTYL_ACCEPTED (actually, PARTLY_ACCEPTED (energy_requested) and DECLINED need to be considered for the last
auction across products, which is not yet implemented).
}
\author{
Sascha Holzhauer
}
