% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_costs.R
\name{hl_figure_energycosts_requested_sumByStartT}
\alias{hl_figure_energycosts_requested_sumByStartT}
\title{Retrieves requests data from DB and creates figure of requested energy and associated costs
summed by delivery start time.}
\usage{
hl_figure_energycosts_requested_sumByStartT(dexpa)
}
\arguments{
\item{dexpa}{parameters}
}
\value{
figure file
}
\description{
TODO consider meter readings
}
\author{
Sascha Holzhauer
}
