% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_costs.R
\name{hl_figure_energycosts_requests_giniByStartT}
\alias{hl_figure_energycosts_requests_giniByStartT}
\title{Retrieves requests data from DB and creates figure of gini coefficient of cost per KWh by delivery start time.}
\usage{
hl_figure_energycosts_requests_giniByStartT(
  dexpa,
  type = "load",
  skiplegend = F
)
}
\arguments{
\item{dexpa}{parameter}

\item{type}{either 'gen' or 'load'}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of gini coefficient of cost per KWh by delivery start time.
}
\author{
Sascha Holzhauer
}
