% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_costs.R
\name{hl_figure_prices_comp_histogram}
\alias{hl_figure_prices_comp_histogram}
\title{Retrieves requests data from DB and creates histogram of prices.}
\usage{
hl_figure_prices_comp_histogram(dexpas, skiplegend = F)
}
\arguments{
\item{dexpa}{parameter}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates histogram of prices.
}
\author{
Sascha Holzhauer
}
