% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_requests.R
\name{hl_figure_requests_numClients_comp_byTypeByStartT}
\alias{hl_figure_requests_numClients_comp_byTypeByStartT}
\title{Retrieves requests data from DB and creates figure of the number of submitting clients per 
request type (Wind/PV/Storage/Load) by delivery start time.}
\usage{
hl_figure_requests_numClients_comp_byTypeByStartT(dexpas, ...)
}
\arguments{
\item{dexpa}{parameter}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of the number of submitting clients per 
request type (Wind/PV/Storage/Load) by delivery start time.
}
\author{
Sascha Holzhauer
}
