% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_requests.R
\name{hl_figure_requests_numRequests_byProductByStartT}
\alias{hl_figure_requests_numRequests_byProductByStartT}
\title{Retrieves requests data from DB and creates figure of the number of received requests per 
product pattern by delivery start time.}
\usage{
hl_figure_requests_numRequests_byProductByStartT(dexpa, ...)
}
\arguments{
\item{dexpa}{parameter}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of the number of received requests per 
product pattern by delivery start time.
}
\author{
Sascha Holzhauer
}
