% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_requests.R
\name{hl_figure_requests_numRequests_byProductBySubmT}
\alias{hl_figure_requests_numRequests_byProductBySubmT}
\title{Retrieves requests data from DB and creates figure of the number of received 
requests per product pattern by submission time.}
\usage{
hl_figure_requests_numRequests_byProductBySubmT(dexpa, ...)
}
\arguments{
\item{dexpa}{parameter}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of the number of received 
requests per product pattern by submission time.
}
\author{
Sascha Holzhauer
}
