% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_figure_requests.R
\name{hl_figure_requests_numRequests_comp_byStatusBySubmT}
\alias{hl_figure_requests_numRequests_comp_byStatusBySubmT}
\title{Retrieves requests data from DB and creates figure of the number of received 
requests per status by submission time.}
\usage{
hl_figure_requests_numRequests_comp_byStatusBySubmT(dexpas, ...)
}
\arguments{
\item{dexpa}{}
}
\value{
figure file
}
\description{
Retrieves requests data from DB and creates figure of the number of received 
requests per status by submission time.
}
\author{
Sascha Holzhauer
}
