% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_opsim.R
\name{hl_opsim}
\alias{hl_opsim}
\title{Run OpSim control manager, schedule service, net simulation, DEX market backend, and EMG clients}
\usage{
hl_opsim(
  dexpa,
  startmanager = T,
  startsservice = T,
  startnetsim = T,
  emptydb = T,
  runemg = T,
  basetime = as.numeric(round(strptime("30/09/19 12:00", "\%d/\%m/\%y \%H:\%M"),
    "mins")) * 1000,
  outputfile = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id,
    ".log", sep = ""),
  shutdown = F
)
}
\arguments{
\item{dexpa}{}
}
\value{
run DEX with OpSIM
}
\description{
Run OpSim control manager, schedule service, net simulation, DEX market backend, and EMG clients
}
\author{
Sascha Holzhauer
}
