% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_opsim.R
\name{hl_opsim_runmanager}
\alias{hl_opsim_runmanager}
\title{Runs OpSim control manager}
\usage{
hl_opsim_runmanager(dexpa, outfileopsim = paste(dexpa$dirs$output$logs,
  dexpa$sim$id, dexpa$opsim$control$logfile, sep = "/"))
}
\arguments{
\item{dexpa}{}

\item{outfileopsim}{}
}
\value{
run OpSim control manager
}
\description{
Runs OpSim control manager
}
\author{
Sascha Holzhauer
}
