% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_opsim.R
\name{hl_opsim_runnetsim}
\alias{hl_opsim_runnetsim}
\title{Run panda power net simulation in python}
\usage{
hl_opsim_runnetsim(dexpa, outfilenetsim = paste(dexpa$dirs$output$logs,
  dexpa$sim$id, dexpa$opsim$netsim$logfile, sep = "/"))
}
\arguments{
\item{dexpa}{}

\item{outfilenetsim}{}
}
\value{
run panda power
}
\description{
Run panda power net simulation in python
}
\author{
Sascha Holzhauer
}
