% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_opsim.R
\name{hl_opsim_runscheduleservice}
\alias{hl_opsim_runscheduleservice}
\title{Run Schedule Service}
\usage{
hl_opsim_runscheduleservice(
  dexpa,
  outfilesservice = paste(dexpa$dirs$output$logs, dexpa$sim$id,
    dexpa$opsim$sservice$logfile, sep = "/")
)
}
\arguments{
\item{dexpa}{}

\item{outfilesservice}{}
}
\value{
run Schedule Service
}
\description{
Run Schedule Service
}
\author{
Sascha Holzhauer
}
