% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_opsim.R
\name{hl_opsim_startmarket}
\alias{hl_opsim_startmarket}
\title{Start market backend server (creating database, not shutting down).}
\usage{
hl_opsim_startmarket(
  dexpa,
  basetime = as.numeric(round(strptime("30/09/19 12:00", "\%d/\%m/\%y \%H:\%M"),
    "mins")) * 1000,
  runemg = T,
  shutdown = F
)
}
\arguments{
\item{dexpa}{}
}
\value{

}
\description{
Start market backend server (creating database, not shutting down).
}
\author{
Sascha Holzhauer
}
