% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_raspic.R
\name{hl_raspic}
\alias{hl_raspic}
\title{Runs experiments on the rasperry Pi cluster}
\usage{
hl_raspic(
  dexpa,
  basetime = as.numeric(round(Sys.time(), "mins")) * 1000,
  offset = round(basetime - as.numeric(Sys.time()) * 1000),
  outputfile = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id, "_",
    dexpa$emg$rseed, ".log", sep = ""),
  outfilemarket = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id,
    "_", dexpa$emg$rseed, "_market.log", sep = ""),
  outfileemg = paste(dexpa$dirs$output$logs, "/", dexpa$sim$id, "/", dexpa$sim$id, "_",
    dexpa$emg$rseed, "_emg.log", sep = "")
)
}
\arguments{
\item{dexpa}{}

\item{basetime}{}

\item{offset}{}

\item{outputfile}{}

\item{outfilemarket}{}

\item{outfileemg}{}
}
\value{
-
}
\description{
Also sets reasonable default values for log files for this application.
}
\author{
Sascha Holzhauer
}
