% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_raspic.R
\name{hl_raspic_closesession}
\alias{hl_raspic_closesession}
\title{Close SSH session to raspberry Pi cluster}
\usage{
hl_raspic_closesession(dexpa, session)
}
\arguments{
\item{dexpa}{}
}
\value{

}
\description{
Close SSH session to raspberry Pi cluster
}
\author{
Sascha Holzhauer

Christoph Hanauer
}
