% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_raspic.R
\name{hl_raspic_runemg}
\alias{hl_raspic_runemg}
\title{Run EMGs on raspberry Pi cluster}
\usage{
hl_raspic_runemg(dexpa, session)
}
\arguments{
\item{dexpa}{}
}
\value{

}
\description{
Run EMGs on raspberry Pi cluster
}
\author{
Sascha Holzhauer

Christoph Hanauer
}
