% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_raspic.R
\name{hl_raspic_transferemgconfig}
\alias{hl_raspic_transferemgconfig}
\title{Transfer config dir to raspberry Pi cluster}
\usage{
hl_raspic_transferemgconfig(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
session
}
\description{
Transfer config dir to raspberry Pi cluster
}
\author{
Sascha Holzhauer

Christoph Hanauer
}
