% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_reports.R
\name{hl_reports_comp}
\alias{hl_reports_comp}
\title{Create a report for the given <code>ids</code> based on the given <code>dexpa</code> (e.g. dumpdir).}
\usage{
hl_reports_comp(
  dexpa,
  ids,
  reporttmpl = "DEX_report_compare_slides.Rmd",
  outfileprefix = paste(dexpa$sim$project, "_ComparisonReport", sep = "_"),
  clean = T,
  dropdb = T
)
}
\arguments{
\item{dexpa}{parameter object required for e.g. dump dir configuration}

\item{ids}{IDs of runs to compare}

\item{reporttmpl}{template for report}

\item{outfileprefix}{prefix for output filename}
}
\value{
comparison report
}
\description{
Loads data from dumpfiles into database and drops these afterwards.
}
\author{
Sascha Holzhauer
}
