% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_statistics.R
\name{hl_statistics_comp_gini}
\alias{hl_statistics_comp_gini}
\title{Table with mean and sum of gini indices across time (load and generation separately)}
\usage{
hl_statistics_comp_gini(dexpas)
}
\arguments{
\item{dexpas}{}
}
\value{

}
\description{
Table with mean and sum of gini indices across time (load and generation separately)
}
\author{
Sascha Holzhauer
}
