% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_statistics.R
\name{hl_statistics_comp_numclients}
\alias{hl_statistics_comp_numclients}
\title{Retrieves request information an outputs aggregated figures of number of clients in a table}
\usage{
hl_statistics_comp_numclients(dexpas)
}
\arguments{
\item{dexpas}{}
}
\value{
table containing aggregated figures
}
\description{
Retrieves request information an outputs aggregated figures of number of clients in a table
}
\author{
Sascha Holzhauer
}
