% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_statistics.R
\name{hl_statistics_energy}
\alias{hl_statistics_energy}
\title{Retrieves clearing and request information an outputs aggregated figures in a table}
\usage{
hl_statistics_energy(dexpa)
}
\arguments{
\item{dexpas}{}
}
\value{
table containing aggregated figures
}
\description{
Retrieves clearing and request information an outputs aggregated figures in a table
}
\author{
Sascha Holzhauer
}
