% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_hl_experiment.R
\name{hl_write_runinfos}
\alias{hl_write_runinfos}
\title{Append current run information to runInfos file.}
\usage{
hl_write_runinfos(dexpa, basetime, offset, infoData)
}
\arguments{
\item{dexpa}{}

\item{basetime}{}

\item{offset}{}

\item{infoData}{}
}
\value{
append runInfos file
}
\description{
Append current run information to runInfos file.
}
\author{
Sascha Holzhauer
}
