% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_csv.R
\name{input_csv_clientdata}
\alias{input_csv_clientdata}
\title{Extract and merge client data from CSV parameter files}
\usage{
input_csv_clientdata(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
data.frame
}
\description{
Extract and merge client data from CSV parameter files
}
\author{
Sascha Holzhauer
}
