% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_csv.R
\name{input_csv_configparam}
\alias{input_csv_configparam}
\title{Reads params for the given dexpa from the parameter configuration CSV file}
\usage{
input_csv_configparam(dexpa, columns = NULL, checkNodeSetId = F)
}
\arguments{
\item{dexpa}{}

\item{columns}{}
}
\value{
vector of columns
}
\description{
Reads params for the given dexpa from the parameter configuration CSV file
}
\author{
Sascha Holzhauer
}
