% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_csv.R
\name{input_csv_runinfos}
\alias{input_csv_runinfos}
\title{Extract the runinfos entry from CSV file for the dexpa's sim ID}
\usage{
input_csv_runinfos(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
data.frame with one row
}
\description{
In case there are more than one entries, return the latest. In case there is none found for the given dexpa ID, the last row
is returned.
}
\author{
Sascha Holzhauer
}
