% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db_clearing.R
\name{input_db_clearings}
\alias{input_db_clearings}
\title{Retrieve clearing data from PostgreSQL database.}
\usage{
input_db_clearings(dexpa)
}
\arguments{
\item{dexpa}{parameter list}
}
\value{
data.frame of clearing infos
}
\description{
Retrieve clearing data from PostgreSQL database.
}
\author{
Sascha Holzhauer
}
