% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_createdb}
\alias{input_db_createdb}
\title{Creates the database configured in the given parameter object}
\usage{
input_db_createdb(dexpa)
}
\arguments{
\item{dexpa}{parameter object}
}
\value{
-
}
\description{
Creates the database configured in the given parameter object
}
\author{
Sascha Holzhauer
}
