% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_db2dump}
\alias{input_db_db2dump}
\title{Exports the database configured in the given parameter object to the given dumpfile}
\usage{
input_db_db2dump(dexpa, dumpdir, remoteServer = FALSE, outputfile = "")
}
\arguments{
\item{dexpa}{parameter object}

\item{dumpdir}{directory to export to (\code{dexpa$dirs$output$dbdumps} is prefixed)}
}
\value{
-
}
\description{
Exports the database configured in the given parameter object to the given dumpfile
}
\author{
Sascha Holzhauer
}
