% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_dropdb}
\alias{input_db_dropdb}
\title{Drop database that is defined in dexpa}
\usage{
input_db_dropdb(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
-
}
\description{
Drop database that is defined in dexpa
}
\author{
Sascha Holzhauer
}
