% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_dropdbs}
\alias{input_db_dropdbs}
\title{Drop databases which are defined in a list of dexpas}
\usage{
input_db_dropdbs(dexpas)
}
\arguments{
\item{dexpas}{list of parameter lists}
}
\value{
-
}
\description{
Drop databases which are defined in a list of dexpas
}
\author{
Sascha Holzhauer
}
