% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_dumps2db}
\alias{input_db_dumps2db}
\title{Import dumpfiles which are defined in a list of dexpas}
\usage{
input_db_dumps2db(dexpas)
}
\arguments{
\item{dexpas}{list of parameter lists}
}
\value{
-
}
\description{
Import dumpfiles which are defined in a list of dexpas
}
\author{
Sascha Holzhauer
}
