% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_getconnection}
\alias{input_db_getconnection}
\title{Retrieve a connection to the PostgreSQL DB specified in passed dexpa}
\usage{
input_db_getconnection(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
database connection
}
\description{
Retrieve a connection to the PostgreSQL DB specified in passed dexpa
}
\author{
Sascha Holzhauer
}
