% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db_parameter.R
\name{input_db_param_products}
\alias{input_db_param_products}
\title{Retrieve market product information from PostgreSQL database.}
\usage{
input_db_param_products(dexpa)
}
\arguments{
\item{dexp}{parameter object}
}
\value{
data.frame of requests
}
\description{
Retrieve market product information from PostgreSQL database.
}
\author{
Sascha Holzhauer
}
