% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db_parameter.R
\name{input_db_param_users}
\alias{input_db_param_users}
\title{Retrieve user parameters from PostgreSQL database.}
\usage{
input_db_param_users(dexp)
}
\arguments{
\item{dexp}{parameter object}
}
\value{
data.frame of requests
}
\description{
Retrieve user parameters from PostgreSQL database.
}
\author{
Sascha Holzhauer
}
