% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db_requests.R
\name{input_db_requests}
\alias{input_db_requests}
\title{Retrieve request data from PostgreSQL database.}
\usage{
input_db_requests(dexpa, additionalwhere = "TRUE")
}
\arguments{
\item{dexp}{parameter object}
}
\value{
data.frame of requests
}
\description{
Retrieve request data from PostgreSQL database.
}
\author{
Sascha Holzhauer
}
