% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db_requests.R
\name{input_db_requests_clearing}
\alias{input_db_requests_clearing}
\title{Retrieve energy requests to illustrate a single clearing process}
\usage{
input_db_requests_clearing(dexpa, starttime = NULL)
}
\arguments{
\item{dexpa}{}

\item{starttime}{time of clearing to be illustrated}
}
\value{
data.frame of energy requests
}
\description{
Retrieve energy requests to illustrate a single clearing process
}
\author{
Sascha Holzhauer
}
