% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_runID}
\alias{input_db_runID}
\title{Generate run ID string of the given database configuration.}
\usage{
input_db_runID(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
run ID string
}
\description{
Generate run ID string of the given database configuration.
}
\author{
Sascha Holzhauer
}
