% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_energy.R
\name{map_requests2intervals_energy}
\alias{map_requests2intervals_energy}
\title{Maps request data to intervals and calculates residuals}
\usage{
map_requests2intervals_energy(dexpa, data, location = F)
}
\arguments{
\item{dexpa}{parameter object}

\item{data}{request data}
}
\value{
data molten data.frame
}
\description{
Maps request data to intervals and calculates residuals
}
\author{
Sascha Holzhauer
}
