% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_clearing.R
\name{output_figure_clearing_comp_numConsideredRequests_byCTbyProduct}
\alias{output_figure_clearing_comp_numConsideredRequests_byCTbyProduct}
\title{Output figure: Number of considered requests per clearing and run ID by clearing time. Adds a legend for run ID.}
\usage{
output_figure_clearing_comp_numConsideredRequests_byCTbyProduct(dexpa,
  data, filename = "dex_clearing_comp_numConsRequestsByCTbyProduct")
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure (file)
}
\description{
Output figure: Number of considered requests per clearing and run ID by clearing time. Adds a legend for run ID.
}
\author{
Sascha Holzhauer
}
