% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_clearing.R
\name{output_figure_clearing_comp_prices_bySTbyProduct}
\alias{output_figure_clearing_comp_prices_bySTbyProduct}
\title{Output figure: Cleared price per clearing by clearing time and product (colour) and run ID.}
\usage{
output_figure_clearing_comp_prices_bySTbyProduct(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Cleared price per clearing by clearing time and product (colour) and run ID.
}
\author{
Sascha Holzhauer
}
