% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_clearing.R
\name{output_figure_clearing_prices_energy_byST}
\alias{output_figure_clearing_prices_energy_byST}
\title{Output figure: Cleared price per clearing by delivery start time.}
\usage{
output_figure_clearing_prices_energy_byST(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure
}
\description{
Output figure: Cleared price per clearing by delivery start time.
}
\author{
Sascha Holzhauer
}
