% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_clientdata.R
\name{output_figure_clientdata}
\alias{output_figure_clientdata}
\title{Prints a bar of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.}
\usage{
output_figure_clientdata(dexpa, d = input_csv_clientdata(dexpa),
  title = NULL, filename = paste(title,
  shbasic::shbasic_condenseRunids(data[, "Runid"]), sep = "_"),
  ggplotaddons = NULL, returnplot = FALSE, facet_ncol = 2)
}
\arguments{
\item{dexpa}{parameter object}

\item{title}{figure title}

\item{filename}{without extension}

\item{ggplotaddons}{vector of ggplot objects to add}

\item{returnplot}{if true the ggplot object is returned}

\item{facet_ncol}{number of columns of facet wrap}

\item{data}{data.frame or list that is rbinded to a data.frame}

\item{y_column}{column name for y data}

\item{x_column}{column name for x data}

\item{fill_column}{column used to define colours}

\item{fill_legendtitle}{title for colour legend}

\item{fill_legenditemnames}{vector of names for colour legend items}

\item{facet_column}{column used to define facets}

\item{alpha}{transparency}

\item{position}{passed to geom_bar}
}
\value{
ggplot2 line visualisation
}
\description{
Prints a bar of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
}
\examples{
print("todo")
}
\author{
Sascha Holzhauer
}
