% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_energy.R
\name{output_figure_energy_requested_byStatusByStartT}
\alias{output_figure_energy_requested_byStatusByStartT}
\title{Output figure: Requested energy per submission time and status.}
\usage{
output_figure_energy_requested_byStatusByStartT(
  dexpa,
  data,
  type = "residual",
  skiplegend = F
)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Requested energy per submission time and status.
}
\author{
Sascha Holzhauer
}
