% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_energy.R
\name{output_figure_energy_requested_comp_sumByStartT}
\alias{output_figure_energy_requested_comp_sumByStartT}
\title{Output figure: Requested energy per submission time and status.}
\usage{
output_figure_energy_requested_comp_sumByStartT(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Requested energy per submission time and status.
}
\author{
Sascha Holzhauer
}
