% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_energy.R
\name{output_figure_energy_requested_sumByLoadGenByStartT}
\alias{output_figure_energy_requested_sumByLoadGenByStartT}
\title{Output figure: Requested energy per submission time and status.}
\usage{
output_figure_energy_requested_sumByLoadGenByStartT(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Requirement: Delivery periods of all products must be a multiple of the shortest delivery period (with start and end
times matching those of the shortest delivery period product)!
}
\seealso{
\code{\link{output_figure_lines}}
}
\author{
Sascha Holzhauer
}
