% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_energy.R
\name{output_figure_energy_requested_sumByStartT}
\alias{output_figure_energy_requested_sumByStartT}
\title{Output figure: Requested energy sum per submission time.}
\usage{
output_figure_energy_requested_sumByStartT(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Requested energy sum per submission time.
}
\author{
Sascha Holzhauer
}
