% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_costs.R
\name{output_figure_energycosts_requested_comp_avgByTypeStartT}
\alias{output_figure_energycosts_requested_comp_avgByTypeStartT}
\title{Output figure: Comparison of average costs of requested energy per type and per delivery start time.}
\usage{
output_figure_energycosts_requested_comp_avgByTypeStartT(
  dexpa,
  data,
  skiplegend = FALSE
)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Comparison of average costs of requested energy per type and per delivery start time.
}
\author{
Sascha Holzhauer
}
