% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_costs.R
\name{output_figure_energycosts_requested_comp_sumByStartT}
\alias{output_figure_energycosts_requested_comp_sumByStartT}
\title{Output figure: Comparison of costs of requested energy sum per delivery start time.}
\usage{
output_figure_energycosts_requested_comp_sumByStartT(dexpa, data,
  skiplegend = F)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Comparison of costs of requested energy sum per delivery start time.
}
\author{
Sascha Holzhauer
}
