% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_costs.R
\name{output_figure_energycosts_requested_giniByStartT}
\alias{output_figure_energycosts_requested_giniByStartT}
\title{Output figure: Gini coefficient of costs of accepted energy per KWh sum per delivery start time.}
\usage{
output_figure_energycosts_requested_giniByStartT(dexpa, data, type,
  skiplegend = F)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Gini coefficient of costs of accepted energy per KWh sum per delivery start time.
}
\author{
Sascha Holzhauer
}
