% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_init.R
\name{output_figure_init}
\alias{output_figure_init}
\title{Init figure output}
\usage{
output_figure_init(
  dexpa,
  outdir = dexpa$dirs$output$figures,
  filename,
  ensurePath = TRUE
)
}
\arguments{
\item{dexpa}{DEX parameter object}

\item{outdir}{(default: dexpa$dirs$output$figures)}

\item{filename}{}

\item{ensurePath}{switch}
}
\value{
plot
}
\description{
Init figure output
}
\author{
Sascha Holzhauer
}
