% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_lines.R
\name{output_figure_lines}
\alias{output_figure_lines}
\title{Prints a lines of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
Uses ticks as X variable if present, runid otherwise.}
\usage{
output_figure_lines(dexpa = dexpa, data = data, x_column = NULL,
  y_column, title = NULL, colour_column = NULL,
  colour_legendtitle = colour_column, colour_legenditemnames = NULL,
  linetype_column = NULL, linetype_legendtitle = linetype_column,
  linetype_legenditemnames = NULL, facet_column = NULL,
  facet_ncol = 2, filename = paste(dexpa$fig$filenameprefix, gsub(" ",
  "_", title), shbasic::shbasic_condenseRunids(data[, "id"]),
  dexpa$fig$filenamepostfix, sep = "_"), alpha = dexpa$fig$alpha,
  showsd = FALSE, ggplotaddons = NULL, returnplot = FALSE)
}
\arguments{
\item{dexpa}{parameter object}

\item{data}{data.frame or list that is rbinded to a data.frame}

\item{x_column}{column name for x data}

\item{y_column}{column name for y data}

\item{title}{figure title}

\item{colour_column}{column used to define colours}

\item{colour_legendtitle}{title for colour legend}

\item{colour_legenditemnames}{vector of names for colour legend items}

\item{linetype_column}{column used to define linetypes}

\item{linetype_legendtitle}{title for linetype legend}

\item{linetype_legenditemnames}{vector of names for linetype legend items}

\item{facet_column}{column used to define facets}

\item{facet_ncol}{number of columns of facet wrap}

\item{filename}{without extension}

\item{alpha}{transparency}

\item{showsd}{if \code{TRUE} line shows mean and ribbons standard deviation (requires multiple data points per group)}

\item{ggplotaddons}{vector of ggplot objects to add}

\item{returnplot}{if true the ggplot object is returned}
}
\value{
ggplot2 line visualisation
}
\description{
Prints a lines of data as ggplot2 with potentially different colour, linetype,
and potentially as facet plot.
Uses ticks as X variable if present, runid otherwise.
}
\seealso{
\code{\link{output_figure_energy_requested_comp_sumByLoadGenByStartT}}
}
\author{
Sascha Holzhauer
}
