% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_energy.R
\name{output_figure_power_status_comp_sumByLocationLoadGenByStartT}
\alias{output_figure_power_status_comp_sumByLocationLoadGenByStartT}
\title{Output figure: Requested energy for specific status per location and submission time.}
\usage{
output_figure_power_status_comp_sumByLocationLoadGenByStartT(dexpa, data,
  status = "REQUESTED", ggplotaddons = NULL)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Requirement: Delivery periods of all products must be a multiple of the shortest delivery period (with start and end
times matching those of the shortest delivery period product)!
}
\seealso{
\code{\link{output_figure_lines}}
}
\author{
Sascha Holzhauer
}
