% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_costs.R
\name{output_figure_prices_comp_histogram}
\alias{output_figure_prices_comp_histogram}
\title{Generate histogram of prices of load requests.}
\usage{
output_figure_prices_comp_histogram(dexpas, data, skiplegend = FALSE,
  title = "Histogram of load prices", ggplotaddons = NULL,
  returnplot = FALSE, filename = paste("Histogram_Load_Prices_",
  shbasic::shbasic_condenseRunids(data[, "id"]), sep = "_"))
}
\arguments{
\item{data}{}

\item{skiplegend}{}
}
\value{
price histogram
}
\description{
Generate histogram of prices of load requests.
}
\author{
Sascha Holzhauer
}
