% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_requests.R
\name{output_figure_requests_numRequests_byStatusByStartT}
\alias{output_figure_requests_numRequests_byStatusByStartT}
\title{Output figure: Number of submitted requests per submission time and status.}
\usage{
output_figure_requests_numRequests_byStatusByStartT(
  dexpa,
  data,
  filename = "dex_requests_numRequests_byStatusByCT"
)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Number of submitted requests per submission time and status.
}
\author{
Sascha Holzhauer
}
