% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_requests.R
\name{output_figure_requests_numRequests_byStatusBySubmT}
\alias{output_figure_requests_numRequests_byStatusBySubmT}
\title{Output figure: Number of considered requests by submission time and status.}
\usage{
output_figure_requests_numRequests_byStatusBySubmT(
  dexpa,
  data,
  subtimerounding = "1 minute",
  filename = "dex_requests_numRequests_byStatusBySubmT"
)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Number of considered requests by submission time and status.
}
\author{
Sascha Holzhauer
}
