% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_requests.R
\name{output_figure_requests_numRequests_comp_byProductByStartT}
\alias{output_figure_requests_numRequests_comp_byProductByStartT}
\title{Output figure: Number of submitted requests per submission time and product.}
\usage{
output_figure_requests_numRequests_comp_byProductByStartT(dexpa, data,
  filename = "dex_requests_numRequests_byProductByCT", skiplegend = F)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Number of submitted requests per submission time and product.
}
\author{
Sascha Holzhauer
}
