% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_requests.R
\name{output_figure_requests_numRequests_comp_byProductBySubmT}
\alias{output_figure_requests_numRequests_comp_byProductBySubmT}
\title{Output figure: Number of considered requests by submission time and product.}
\usage{
output_figure_requests_numRequests_comp_byProductBySubmT(
  dexpa,
  data,
  filename = "dex_requests_numRequests_byProductBySubmT",
  skiplegend = F
)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Number of considered requests by submission time and product.
}
\author{
Sascha Holzhauer
}
