% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_requests.R
\name{output_figure_requests_numRequests_comp_byTypeByStartT}
\alias{output_figure_requests_numRequests_comp_byTypeByStartT}
\title{Output figure: Number of submitted requests per submission time and request type.}
\usage{
output_figure_requests_numRequests_comp_byTypeByStartT(dexpa, data,
  filename = "dex_requests_numRequests_byTypeByCT", skiplegend = F)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
figure file
}
\description{
Output figure: Number of submitted requests per submission time and request type.
}
\author{
Sascha Holzhauer
}
