% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_singleclearing.R
\name{output_figure_singleclearing_en}
\alias{output_figure_singleclearing_en}
\title{Plots demand and supply curve for a single uniform price clearing with energy on x-axis (English version)}
\usage{
output_figure_singleclearing_en(dexpa, title = "", offers_df, bids_df)
}
\arguments{
\item{title}{Title to be plotted in figure}

\item{offers_df}{}

\item{bids_df}{}
}
\value{

}
\description{
Plots demand and supply curve for a single uniform price clearing with energy on x-axis (English version)
}
\author{
Sascha Holzhauer
}
