% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_figure_singleclearing.R
\name{output_figure_singleclearing_twisted}
\alias{output_figure_singleclearing_twisted}
\title{Plots demand and supply curve for a single uniform price clearing with price on x-axis}
\usage{
output_figure_singleclearing_twisted(
  dexpa,
  title = "",
  offers_df,
  bids_df,
  offers_comp = NULL,
  bids_comp = NULL
)
}
\arguments{
\item{title}{Title to be plotted in figure}

\item{offers_df}{}

\item{bids_df}{}
}
\value{
plot
}
\description{
Plots demand and supply curve for a single uniform price clearing with price on x-axis
}
\author{
Friedrich Krebs

Sascha Holzhauer
}
