% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_statistics.R
\name{output_statistics_comp_costs}
\alias{output_statistics_comp_costs}
\title{Calculate aggregate figures of costs.}
\usage{
output_statistics_comp_costs(dexpa, cinfos, requestdata)
}
\arguments{
\item{dexpa}{}

\item{cinfos}{}
}
\value{
markdown table
}
\description{
Calculate aggregate figures of costs.
}
\author{
Sascha Holzhauer
}
