% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_statistics.R
\name{output_statistics_comp_energy}
\alias{output_statistics_comp_energy}
\title{Calculate aggregate figures of traded energy.}
\usage{
output_statistics_comp_energy(dexpa, cinfos, requestdata)
}
\arguments{
\item{dexpa}{}

\item{cinfos}{}
}
\value{
markdown table
}
\description{
Calculate aggregate figures of traded energy.
}
\author{
Sascha Holzhauer
}
