% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_statistics.R
\name{output_statistics_comp_gini}
\alias{output_statistics_comp_gini}
\title{Show aggregate gini figures in table per ID.}
\usage{
output_statistics_comp_gini(dexpas, data)
}
\arguments{
\item{data}{data.frame with column "id' and individual data point in column 'gini'}

\item{dexpa}{}
}
\value{
markdown table
}
\description{
Show aggregate gini figures in table per ID.
}
\author{
Sascha Holzhauer
}
