% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_statistics.R
\name{output_statistics_comp_numclients}
\alias{output_statistics_comp_numclients}
\title{Calculate aggregate figures of number of clients.}
\usage{
output_statistics_comp_numclients(dexpa, requestdata)
}
\arguments{
\item{dexpa}{}

\item{requestdata}{}
}
\value{
markdown table
}
\description{
Calculate aggregate figures of number of clients.
}
\author{
Sascha Holzhauer
}
