% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_table_param.R
\name{output_table_param_clients}
\alias{output_table_param_clients}
\title{Returns a kable markdown table with client information}
\usage{
output_table_param_clients(
  dexpa,
  format = "markdown",
  caption = "Client information",
  linespertable = 18,
  file = NULL,
  filextension = "tex",
  prefix = ""
)
}
\arguments{
\item{dexpa}{}
}
\value{
kable markdown table
}
\description{
Returns a kable markdown table with client information
}
\author{
Sascha Holzhauer
}
