% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_table_param.R
\name{output_table_param_marketinfo}
\alias{output_table_param_marketinfo}
\title{Provides markdown formated table of timing information.}
\usage{
output_table_param_marketinfo(dexpa, format = "markdown",
  caption = "Market information")
}
\arguments{
\item{dexp}{}
}
\value{
table of timing information
}
\description{
Provides markdown formated table of timing information.
}
\author{
Sascha Holzhauer
}
