% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_table_param.R
\name{output_table_param_products}
\alias{output_table_param_products}
\title{Provides markdown formated table of product patterns.}
\usage{
output_table_param_products(dexpas, format = "markdown",
  caption = "Product information", skipcolumns = c())
}
\arguments{
\item{dexp}{}
}
\value{
table of product patterns
}
\description{
Provides markdown formated table of product patterns.
}
\author{
Sascha Holzhauer
}
