% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{param_clientsdiffer}
\alias{param_clientsdiffer}
\title{Check differences in client configuration (compare rows of configuration belonging to same ID)}
\usage{
param_clientsdiffer(dexpa)
}
\value{
true if client configurations differ between rows of passed dexpa
}
\description{
Check differences in client configuration (compare rows of configuration belonging to same ID)
}
\author{
Sascha Holzhauer
}
