% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{param_clientsdiffer_runs}
\alias{param_clientsdiffer_runs}
\title{Check differences in client configuration (compare runs)}
\usage{
param_clientsdiffer_runs(dexpas)
}
\value{
true if client configurations differ between passed dexpas
}
\description{
Check differences in client configuration (compare runs)
}
\author{
Sascha Holzhauer
}
