% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_params.R
\name{param_getDefaultDexpa}
\alias{param_getDefaultDexpa}
\title{Get the default dexpa parameter list}
\usage{
param_getDefaultDexpa()
}
\value{
list of default dexpa parameters
}
\description{
Get the default dexpa parameter list
}
\author{
Sascha Holzhauer
}
