% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{param_marketinfodiffer}
\alias{param_marketinfodiffer}
\title{Check differences in market information}
\usage{
param_marketinfodiffer(dexpas)
}
\arguments{
\item{dexpas}{}
}
\value{
true if market information differ between passed dexpas
}
\description{
Check differences in market information
}
\author{
Sascha Holzhauer
}
