% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{param_marketproductsdiffer}
\alias{param_marketproductsdiffer}
\title{Check differences in market product configuration}
\usage{
param_marketproductsdiffer(dexpas)
}
\arguments{
\item{dexpas}{}
}
\value{
true if market product configurations differ between passed dexpas
}
\description{
Check differences in market product configuration
}
\author{
Sascha Holzhauer
}
