% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_params.R
\name{param_mergeDefaultDexpa}
\alias{param_mergeDefaultDexpa}
\title{Merge the given list into the default dexpa parameter list.
Applies \code{\link[utils]{mergeList}}.}
\usage{
param_mergeDefaultDexpa(dexpa = list())
}
\arguments{
\item{dexpa}{}
}
\value{
given list modified by adding undefined parameter of dexpa default parameter list.
}
\description{
Merge the given list into the default dexpa parameter list.
Applies \code{\link[utils]{mergeList}}.
}
\author{
Sascha Holzhauer
}
