% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_costs.R
\name{prepare_costdata_gini}
\alias{prepare_costdata_gini}
\title{Calculate gini data}
\usage{
prepare_costdata_gini(dexpas, data, type, normalise = F)
}
\arguments{
\item{dexpas}{}

\item{data}{}

\item{type}{either "load" or "gen"}
}
\value{
data
}
\description{
Calculate gini data
}
\author{
Sascha Holzhauer
}
