% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_energy.R,
%   R/dexr_output_figure_energy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%within\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lubridate}{\code{\link[lubridate]{\%within\%}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

