% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_output_reports.R
\name{reports_level_medium}
\alias{reports_level_medium}
\title{Set reporting of figures to medium level}
\usage{
reports_level_medium(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
dexpa
}
\description{
Set reporting of figures to medium level
}
\author{
Sascha Holzhauer
}
