% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_requests.R
\name{requests_filter_data}
\alias{requests_filter_data}
\title{Filter requests according to status and products}
\usage{
requests_filter_data(dexpa, d)
}
\arguments{
\item{dexpa}{}

\item{d}{}
}
\value{

}
\description{
Filter requests according to status and products
}
\author{
Sascha Holzhauer
}
