% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{requests_identify_type}
\alias{requests_identify_type}
\title{Sums the expression in <code>dataexp</code> up separately for PC/Win/StorageOut/StorageIn}
\usage{
requests_identify_type(data,
  dataexp = "df[r, if(df$status==2) \\"energy_accepted\\" else \\"energy_requested\\"]")
}
\arguments{
\item{data}{}

\item{dataexp}{}
}
\value{

}
\description{
Sums the expression in <code>dataexp</code> up separately for PC/Win/StorageOut/StorageIn
}
\author{
Sascha Holzhauer
}
