% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper_costs.R
\name{requests_prices_identify_type}
\alias{requests_prices_identify_type}
\title{Add a column with generation (Pv, Wind, StorageOut) and load type (StorageIn) per Id.}
\usage{
requests_prices_identify_type(dexpa, data)
}
\arguments{
\item{dexpa}{}

\item{data}{}
}
\value{
data
}
\description{
Add a column with generation (Pv, Wind, StorageOut) and load type (StorageIn) per Id.
}
\author{
Sascha Holzhauer
}
