% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_server_run.R
\name{server_isrunning}
\alias{server_isrunning}
\title{Checks whether the server is ready for requests.}
\usage{
server_isrunning(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
TRUE if the server is up
}
\description{
Checks whether the server is ready for requests.
}
\author{
Sascha Holzhauer
}
