% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_server_run.R
\name{server_shutdown}
\alias{server_shutdown}
\title{Stop backend server}
\usage{
server_shutdown(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
shutdown message
}
\description{
Stop backend server
}
\author{
Sascha Holzhauer
}
