% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_server_run.R
\name{server_start}
\alias{server_start}
\title{Start backend server}
\usage{
server_start(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
status code
}
\description{
Start backend server
}
\author{
Sascha Holzhauer
}
