% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_server_run.R
\name{server_status}
\alias{server_status}
\title{Retrieve backend server status information}
\usage{
server_status(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
JSON info map
}
\description{
Retrieve backend server status information
}
\author{
Sascha Holzhauer
}
