% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_server_run.R
\name{server_version}
\alias{server_version}
\title{Retrieve backend server version}
\usage{
server_version(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
JSON info map
}
\description{
Retrieve backend server version
}
\author{
Sascha Holzhauer
}
