% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{setup_environment}
\alias{setup_environment}
\title{Copies machine-specific and project-specific config file <code>dexpa-machine_machine.R</code> to location 
<code>dexpa$dirs$config</code> for the user to edit. Creates folder if not existing.}
\usage{
setup_environment(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
-
}
\description{
Copies machine-specific and project-specific config file <code>dexpa-machine_machine.R</code> to location 
<code>dexpa$dirs$config</code> for the user to edit. Creates folder if not existing.
}
\author{
Sascha Holzhauer
}
