% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{setup_project_version}
\alias{setup_project_version}
\title{Copies config files <code>dexpa-pversion_pversion</code>, run file (<code>DEX_Runs.csv</code>), and 
config table files (<code>DEX_Param_Configs.xxx</code>) to a new project version dir for the user to edit. Given
<code>dexpa</code> needs to have the project version assigned.}
\usage{
setup_project_version(dexpa)
}
\arguments{
\item{dexpa}{}
}
\value{
dexpa
}
\description{
NOTE: The machine-specific config file should have been executed before running this procedure!
}
\details{
NOTE: The returned dexpa object needs to be assigned back since the source procedure operates locally within this function.
}
\author{
Sascha Holzhauer
}
