% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_helper.R
\name{str_to_ms}
\alias{str_to_ms}
\title{Convert time string to milliseconds}
\usage{
str_to_ms(timestring, format = "\%Y-\%m-\%d \%H:\%M",
  timezone = "Europe/Berlin")
}
\arguments{
\item{timestring}{}

\item{format}{}

\item{timezone}{}
}
\value{
time in ms
}
\description{
Convert time string to milliseconds
}
\author{
Sascha Holzhauer
}
