% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexr_input_db.R
\name{input_db_dump2db}
\alias{input_db_dump2db}
\title{Imports the given dumpfile into the database configured in the given parameter object}
\usage{
input_db_dump2db(dexpa, dumpfile = paste("dump_", dexpa$sim$id, sep =
  ""))
}
\arguments{
\item{dexpa}{parameter object}

\item{dumpfile}{dumpfile to import}
}
\value{
-
}
\description{
Imports the given dumpfile into the database configured in the given parameter object
}
\author{
Sascha Holzhauer
}
